+++ packages/atomic/src/components.d.ts
@@ -100,13 +100,13 @@ export namespace Components {
     }
     interface AtomicResultsPerPage {
         /**
-          * Initial value of the result per page option
+          * List of possible results per page choices, separated by commas.
          */
-        "initialOption": number;
+        "choicesDisplayed": string;
         /**
-          * List of possible results per page options, separated by commas
+          * Initial choice for the number of result per page. Should be part of the `choicesDisplayed` option.
          */
-        "options": string;
+        "initialChoice": number;
     }
     interface AtomicSearchBox {
         "_id": string;
@@ -428,13 +428,13 @@ declare namespace LocalJSX {
     }
     interface AtomicResultsPerPage {
         /**
-          * Initial value of the result per page option
+          * List of possible results per page choices, separated by commas.
          */
-        "initialOption"?: number;
+        "choicesDisplayed"?: string;
         /**
-          * List of possible results per page options, separated by commas
+          * Initial choice for the number of result per page. Should be part of the `choicesDisplayed` option.
          */
-        "options"?: string;
+        "initialChoice"?: number;
     }
     interface AtomicSearchBox {
         "_id"?: string;